Refactor BsCustomDates interface out in a separate file bs-customdates.ts
