+++ src/datepicker/base/bs-datepicker-container.ts
@@ -1,6 +1,6 @@
 // datepicker container component
 /* tslint:disable:no-empty */
-import { BsCustomDates } from '../themes/bs/bs-custom-dates-view.component';
+import { BsCustomDates } from '../bs-customdates';
 import { BsDatepickerEffects } from '../reducer/bs-datepicker.effects';
 import { Observable } from 'rxjs';
 import {
