+++ src/datepicker/bs-customdates.ts
@@ -1,0 +1,4 @@
+export interface BsCustomDates {
+  label: string;
+  value: Date | Date[];
+}
