+++ src/datepicker/themes/bs/bs-datepicker-container.component.ts
@@ -7,7 +7,7 @@ import { BsDatepickerActions } from '../../reducer/bs-datepicker.actions';
 import { BsDatepickerEffects } from '../../reducer/bs-datepicker.effects';
 import { BsDatepickerStore } from '../../reducer/bs-datepicker.store';
 import { Subscription } from 'rxjs';
-import { BsCustomDates } from './bs-custom-dates-view.component';
+import { BsCustomDates } from '../../bs-customdates';
 
 @Component({
   selector: 'bs-datepicker-container',
